Changes along the way of moving to formly forms. Fiddling with promises and some bug fixes.
Updated dependencies yielding bower.json changes.
Changed addProject view to a formly form.
Changed the Description detail view to a formly form. In the project
view the detail viewport was inside a form, and now it is in the
viewport.
Fiddling with onEnter and resolve entries in the two config files. Still
not happy. The select state makes a call to the back end for a pristine,
empty description model, which is used in the addProject view. Trying to
supply that with a resolve in the config file. There are corresponding
changes in the controllers.
Lots of whitespace changes in projectDataService. I tried using JSLint
and was trying to quiet it down. Some new methods.
Some real bugs fixed in projectListService. Reworked the setProjectID
method.