Much progress with ui.router resolve. Not throwing away promises any more.
Using the promises from $http and not instantiating my own.
Starting to have modules be the lowest level to be in a closure, instead
of file level. I like getting two columns back.
Moved all of the formly configuration to the attributes module, which
should be cleaned up and made a "formly" module.
Still one ui.router "onEnter" to get rid of, in the filter module.