Updating the description table with a formly form is working.
Moved the formly configuration and custom types definitions into a
(properly configured) attributesConfig file. Date pickers are working.
The big effort of the day was the unsaved changes handling when
switching between tabs under the project.edit state: getting unwanted
changes to be thrown away, getting input to make the form dirty again,
and marking the form pristine after saving and updating the form.
Moved the checks for unsaved changes in the project update form into the
projectDataService. The addProject form is not done.
White space, and comment changes sprinkled about. Got rid of a few more
enclosures.