More followup to UI-Router update.
Done with the stateLocation module. UI-Router update made it obsolete.
Took out references to ui.router.state.events, inserted to get things
working after the update.
Done with the loginInjector module. Will use $transitions.onBefore to
check for permission.
Introduced $transitions to eliminate $stateChange... event watchers.They
are all gone.
Introduced $transitions$ into config files to move away from
$stateParams.

Replaced the viewport on the index page with two named viewports. The
one named "app" is for a tree of all the states under all of the tabs.
Created a new top level "app" state, and put existing select, project,
... states under it. This required updating all the config files, and
changes to navigation from tab to tab. The app state is defined in the
top level app.config file.

The new viewport is named "modal", and allows for modal states. I need
to be able to use resolve to ensure data items are present to modals,
like the csrf token required to login. I envision moving login and
unsaved changes checks to new states under "modal", and perhaps adding a
modal.user state for preferences. Rewriting login functionality is the
next task.

Revised report config with resolves to ensure that the
projectListService and the reportTableService are properly initialized.
Got rid of the service init function in reportTableService that should
never have been there.
Fixed a bug in projectListService that was not selecting the right
project name if you jumped into the middle of a list of projects or used
a filter to select a subset of all projects. Needed a better resolve in
projectConfig and small changes in projectListService.